+++ src/hosted.rs
@@ -4,7 +4,7 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-const BUFFER_SIZE_LIMIT: usize = 1024 * 2;
+const BUFFER_SIZE_LIMIT: usize = 1024 * 16;
 
 use canonical::{BridgeStore, ByteSink, ByteSource, Canon, Id32, Store};
 
@@ -99,14 +99,23 @@ where
 }
 
 /// Call another contract at address `target`
-pub fn transact_raw(
+pub fn transact_raw<S, Slf>(
+    slf: &mut Slf,
     target: &ContractId,
     transaction: &Transaction,
-) -> Result<(ContractState, ReturnValue), <BridgeStore<Id32> as Store>::Error> {
-    let bs = BridgeStore::<Id32>::default();
+) -> Result<ReturnValue, S::Error>
+where
+    S: Store,
+    Slf: Canon<S>,
+{
+    let bs = S::default();
 
     let mut buf = [0u8; BUFFER_SIZE_LIMIT];
     let mut sink = ByteSink::new(&mut buf, &bs);
+
+    // Store the state before call `transact`
+    let state = ContractState::from_canon(slf, &bs)?;
+    state.write(&mut sink)?;
 
     transaction.write(&mut sink)?;
 
@@ -115,31 +124,31 @@ pub fn transact_raw(
     // read return back
     let mut source = ByteSource::new(&buf, &bs);
 
-    Ok((
-        ContractState::read(&mut source)?,
-        ReturnValue::read(&mut source)?,
-    ))
+    let state = ContractState::read(&mut source)?;
+
+    *slf = state.cast(bs.clone())?;
+
+    Ok(ReturnValue::read(&mut source)?)
 }
 
 /// Call another contract at address `target`
 ///
 /// Note that you will have to specify the expected return and argument types
 /// yourself.
-pub fn transact<A, R, Slf>(
+pub fn transact<S, A, R, Slf>(
     slf: &mut Slf,
     target: &ContractId,
     transaction: &A,
-) -> Result<R, <BridgeStore<Id32> as Store>::Error>
+) -> Result<R, S::Error>
 where
-    A: Canon<BridgeStore<Id32>>,
-    R: Canon<BridgeStore<Id32>>,
-    Slf: Canon<BridgeStore<Id32>>,
+    S: Store,
+    A: Canon<S>,
+    R: Canon<S>,
+    Slf: Canon<S>,
 {
-    let bs = BridgeStore::<Id32>::default();
+    let bs = S::default();
     let wrapped = Transaction::from_canon(transaction, &bs)?;
-    let (state, result) = transact_raw(&target, &wrapped)?;
-
-    *slf = state.cast(bs)?;
+    let result = transact_raw(slf, &target, &wrapped)?;
 
     result.cast(bs)
 }